requirements.txt
google-api-python-client>=2.86.0
google-auth>=2.17.3
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
docx2txt>=0.8
requests>=2.28.2
pydantic>=2.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
matplotlib>=3.4.0
argparse>=1.4.0
pyautogen>=0.2.0